Issue 21826: Move some classes around to avoid package resolution errors
Had to move some classes around to avoid the following:

```
[ERROR   ] CWWKE0702E: Could not resolve module: com.ibm.ws.security.openidconnect.common [149]
  Bundle was not resolved because of a uses constraint violation.
  org.apache.felix.resolver.reason.ReasonException: Uses constraint violation. Unable to resolve resource com.ibm.ws.security.openidconnect.common [osgi.identity; osgi.identity="com.ibm.ws.security.openidconnect.common"; type="osgi.bundle"; version:Version="1.0.77.202304261135"] because it exports package 'com.ibm.ws.security.openidconnect.common' and is also exposed to it from resource com.ibm.ws.security.openidconnect.clients.common [osgi.identity; osgi.identity="com.ibm.ws.security.openidconnect.clients.common"; type="osgi.bundle"; version:Version="1.0.77.202304261135"] via the following dependency chain:
```

My guess is that opening up the `com.ibm.ws.security.openidconnect.clients.common` bundle that had its own `com.ibm.ws.security.openidconnect.common` package conflicted with the actual `com.ibm.ws.security.openidconnect.common` project.
