+++ dev/com.ibm.ws.security.openidconnect.clients.common/src/com/ibm/ws/security/openidconnect/clients/common/AuthorizationCodeHandler.java
@@ -27,8 +27,8 @@ import com.ibm.websphere.ssl.SSLException;
 import com.ibm.ws.security.common.ssl.NoSSLSocketFactoryException;
 import com.ibm.ws.security.common.structures.BoundedHashMap;
 import com.ibm.ws.security.openidconnect.client.jose4j.util.Jose4jUtil;
+import com.ibm.ws.security.openidconnect.clients.common.token.auth.TokenEndpointAuthMethod;
 import com.ibm.ws.security.openidconnect.pkce.ProofKeyForCodeExchangeHelper;
-import com.ibm.ws.security.openidconnect.common.token.auth.TokenEndpointAuthMethod;
 import com.ibm.ws.webcontainer.security.AuthResult;
 import com.ibm.ws.webcontainer.security.ProviderAuthenticationResult;
 import com.ibm.wsspi.ssl.SSLSupport;
