+++ dev/com.ibm.ws.security.openidconnect.clients.common/src/com/ibm/ws/security/openidconnect/clients/common/TraceConstantsCommonMessages.java
@@ -1,18 +1,16 @@
 /*******************************************************************************
- * Copyright (c) 2011, 2018 IBM Corporation and others.
+ * Copyright (c) 2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
- * SPDX-License-Identifier: EPL-2.0
  *
- * Contributors:
- *     IBM Corporation - initial API and implementation
+ * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package com.ibm.ws.security.openidconnect.common;
+package com.ibm.ws.security.openidconnect.clients.common;
 
-public interface TraceConstants {
-    public final static String TRACE_GROUP = "OpenIdConnect";
+public interface TraceConstantsCommonMessages {
+
+    public final static String TRACE_GROUP = "OPENIDCONNECT";
     public final static String MESSAGE_BUNDLE = "com.ibm.ws.security.openidconnect.common.internal.resources.OidcCommonMessages";
 }
