+++ dev/com.ibm.ws.security.openidconnect.common/src/com/ibm/websphere/security/openidconnect/package-info.java
@@ -1,21 +1,18 @@
 /*******************************************************************************
- * Copyright (c) 2013 IBM Corporation and others.
+ * Copyright (c) 2013, 2023 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
- * SPDX-License-Identifier: EPL-2.0
  *
- * Contributors:
- *     IBM Corporation - initial API and implementation
+ * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
 /**
  * @version 1.0.0
  */
 @org.osgi.annotation.versioning.Version("1.0.0")
-@TraceOptions(traceGroup = "OpenIdConnect", messageBundle = "com.ibm.ws.security.openidconnect.common.internal.resources.OidcCommonMessages")
+@TraceOptions(traceGroup = TraceConstantsCommonMessages.TRACE_GROUP, messageBundle = TraceConstantsCommonMessages.MESSAGE_BUNDLE)
 package com.ibm.websphere.security.openidconnect;
 
 import com.ibm.websphere.ras.annotation.TraceOptions;
-
+import com.ibm.ws.security.openidconnect.clients.common.TraceConstantsCommonMessages;
