+++ dev/com.ibm.ws.security.openidconnect.server/resources/sessionMgmt.jsp
@@ -1,6 +1,6 @@
 <%@ page language="java" contentType="text/html; charset=utf-8"
     pageEncoding="utf-8"
-	import="com.ibm.ws.security.openidconnect.common.Constants"
+	import="com.ibm.ws.security.openidconnect.clients.common.Constants"
 	import="java.net.URL"
 %>
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
